INT-2829 Remove TCP Package Tangle
tcp.connection <-> tcp.connection.support

Remove ...connection.support package - move classes
to ...connection.
