--- spring-integration-ip/src/main/java/org/springframework/integration/ip/tcp/connection/support/package-info.java
@@ -1,5 +1,0 @@
-/**
- * Provides classes supporting the creation/manipulation of sockets,
- * SSLContexts etc.
- */
-package org.springframework.integration.ip.tcp.connection.support;
