+++ spring-integration-ip/src/test/java/org/springframework/integration/ip/tcp/connection/SocketSupportTests.java
@@ -40,11 +40,6 @@ import org.mockito.Mockito;
 import org.mockito.invocation.InvocationOnMock;
 import org.mockito.stubbing.Answer;
 import org.springframework.integration.Message;
-import org.springframework.integration.ip.tcp.connection.support.DefaultTcpNetSSLSocketFactorySupport;
-import org.springframework.integration.ip.tcp.connection.support.DefaultTcpSSLContextSupport;
-import org.springframework.integration.ip.tcp.connection.support.TcpSSLContextSupport;
-import org.springframework.integration.ip.tcp.connection.support.TcpSocketFactorySupport;
-import org.springframework.integration.ip.tcp.connection.support.TcpSocketSupport;
 import org.springframework.integration.ip.tcp.serializer.ByteArrayCrLfSerializer;
 import org.springframework.integration.ip.util.TestingUtilities;
 import org.springframework.integration.message.GenericMessage;
